pyproject.toml
[tool.poetry]
name = "dbt_faker"
version = "0.1.0"
description = "Fake data generator for dbt demos"
authors = ["Doug Guthrie <douglas.p.guthrie@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
factory-boy = "^3.2.1"
snowflake-sqlalchemy = "^1.3.4"
pandas = "^1.4.2"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pre-commit = "^2.19.0"
isort = "^5.10.1"
mypy = "^0.961"
flake8 = "^4.0.1"
pytest = "^7.1.2"
ipython = "^8.4.0"
coverage = "^6.4.1"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
